refactor: Gool 컴포넌트에서 사용하는 Background, ClickButton 컴포넌트로 분리
